requirements.txt
pymysql>=1.1.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pytest>=8.3.0
snowflake-connector-python[pandas]>=3.0.0
pandas>=2.0.0
